Remove: EditorMotionDispatcher
In order to implement fallback during edit mode, schedule processing on the Editor is unified to the MotionDispatcher class.
(This is a breaking change.)
