--- src/LitMotion/Assets/LitMotion/Editor/EditorMotionDispatcher.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9c738daa6ff2b47f5b1130a82084b0d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
