+++ src/LitMotion/Assets/LitMotion/Editor/EditorUpdateMotionScheduler.cs
@@ -7,7 +7,7 @@ namespace LitMotion.Editor
             where TOptions : unmanaged, IMotionOptions
             where TAdapter : unmanaged, IMotionAdapter<TValue, TOptions>
         {
-            return EditorMotionDispatcher.Schedule<TValue, TOptions, TAdapter>(data, callbackData);
+            return MotionDispatcher.Schedule<TValue, TOptions, TAdapter>(data, callbackData, UpdateMode.EditorApplicationUpdate);
         }
     }
 }
