+++ src/LitMotion/Assets/LitMotion/Runtime/MotionDispatcher.cs
@@ -1,12 +1,18 @@
 using UnityEngine;
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
 
 namespace LitMotion
 {
     internal enum UpdateMode
     {
-        Update,
-        LateUpdate,
-        FixedUpdate
+        Update = 0,
+        LateUpdate = 1,
+        FixedUpdate = 2,
+#if UNITY_EDITOR
+        EditorApplicationUpdate = 10
+#endif
     }
 
     /// <summary>
@@ -27,6 +33,9 @@ namespace LitMotion
             public static MotionStorage<TValue, TOptions, TAdapter> update;
             public static MotionStorage<TValue, TOptions, TAdapter> lateUpdate;
             public static MotionStorage<TValue, TOptions, TAdapter> fixedUpdate;
+#if UNITY_EDITOR
+            public static MotionStorage<TValue, TOptions, TAdapter> editorApplicationUpdate;
+#endif
 
             public static MotionStorage<TValue, TOptions, TAdapter> GetOrCreate(UpdateMode updateMode)
             {
@@ -57,6 +66,16 @@ namespace LitMotion
                             fixedUpdate = storage;
                         }
                         return fixedUpdate;
+#if UNITY_EDITOR
+                    case UpdateMode.EditorApplicationUpdate:
+                        if (editorApplicationUpdate == null)
+                        {
+                            var storage = new MotionStorage<TValue, TOptions, TAdapter>(MotionStorageManager.CurrentStorageId);
+                            MotionStorageManager.AddStorage(storage);
+                            editorApplicationUpdate = storage;
+                        }
+                        return editorApplicationUpdate;
+#endif
                 }
             }
         }
@@ -69,12 +88,18 @@ namespace LitMotion
             public static UpdateRunner<TValue, TOptions, TAdapter> update;
             public static UpdateRunner<TValue, TOptions, TAdapter> lateUpdate;
             public static UpdateRunner<TValue, TOptions, TAdapter> fixedUpdate;
+#if UNITY_EDITOR
+            public static UpdateRunner<TValue, TOptions, TAdapter> editorApplicationUpdate;
+#endif
         }
 
         static readonly MinimumList<IUpdateRunner> updateRunners = new();
         static readonly MinimumList<IUpdateRunner> lateUpdateRunners = new();
         static readonly MinimumList<IUpdateRunner> fixedUpdateRunners = new();
-        
+#if UNITY_EDITOR
+        static readonly MinimumList<IUpdateRunner> editorApplicationUpdateRunners = new();
+#endif
+
         /// <summary>
         /// Ensures the storage capacity until it reaches at least `capacity`.
         /// </summary>
@@ -87,12 +112,15 @@ namespace LitMotion
             StorageCache<TValue, TOptions, TAdapter>.GetOrCreate(UpdateMode.Update).EnsureCapacity(capacity);
             StorageCache<TValue, TOptions, TAdapter>.GetOrCreate(UpdateMode.LateUpdate).EnsureCapacity(capacity);
             StorageCache<TValue, TOptions, TAdapter>.GetOrCreate(UpdateMode.FixedUpdate).EnsureCapacity(capacity);
+#if UNITY_EDITOR
+            StorageCache<TValue, TOptions, TAdapter>.GetOrCreate(UpdateMode.EditorApplicationUpdate).EnsureCapacity(capacity);
+#endif
         }
 
         internal static MotionHandle Schedule<TValue, TOptions, TAdapter>(in MotionData<TValue, TOptions> data, in MotionCallbackData callbackData, UpdateMode updateMode)
-           where TValue : unmanaged
-           where TOptions : unmanaged, IMotionOptions
-           where TAdapter : unmanaged, IMotionAdapter<TValue, TOptions>
+            where TValue : unmanaged
+            where TOptions : unmanaged, IMotionOptions
+            where TAdapter : unmanaged, IMotionAdapter<TValue, TOptions>
         {
             MotionStorage<TValue, TOptions, TAdapter> storage = StorageCache<TValue, TOptions, TAdapter>.GetOrCreate(updateMode);
             switch (updateMode)
@@ -122,6 +150,16 @@ namespace LitMotion
                         RunnerCache<TValue, TOptions, TAdapter>.fixedUpdate = runner;
                     }
                     break;
+#if UNITY_EDITOR
+                case UpdateMode.EditorApplicationUpdate:
+                    if (RunnerCache<TValue, TOptions, TAdapter>.editorApplicationUpdate == null)
+                    {
+                        var runner = new UpdateRunner<TValue, TOptions, TAdapter>(storage);
+                        editorApplicationUpdateRunners.Add(runner);
+                        RunnerCache<TValue, TOptions, TAdapter>.editorApplicationUpdate = runner;
+                    }
+                    break;
+#endif
             }
 
             var (EntryIndex, Version) = storage.Append(data, callbackData);
@@ -174,5 +212,27 @@ namespace LitMotion
                 array[i]?.Reset();
             }
         }
+
+#if UNITY_EDITOR
+        static double lastEditorTime;
+
+        [InitializeOnLoadMethod]
+        static void InitEditor()
+        {
+            lastEditorTime = 0f;
+            EditorApplication.update += UpdateEditor;
+        }
+
+        static void UpdateEditor()
+        {
+            var deltaTime = (float)(EditorApplication.timeSinceStartup - lastEditorTime);
+            var array = updateRunners.AsArray();
+            for (int i = 0; i < array.Length; i++)
+            {
+                array[i]?.Update(deltaTime, deltaTime);
+            }
+            lastEditorTime = EditorApplication.timeSinceStartup;
+        }
+#endif
     }
 }
